multi-test fixes: whitespace in wolfcrypt/src/random.c and wolfcrypt/test/test.c, bugprone-macro-parentheses and -Wenum-compare in WS_RETURN_CODE() (wolfssl/ssl.h), and clang-analyzer-deadcode.DeadStores in api.c.

+++ wolfssl/ssl.h
@@ -786,7 +786,7 @@ enum SNICbReturn {
  */
 #if defined(WOLFSSL_ERROR_CODE_OPENSSL)
     #define WS_RETURN_CODE(item1,item2) \
-      ((item1 < 0) ? item2 : item1)
+      (((item1) < 0) ? (int)(item2) : (int)(item1))
 #else
     #define WS_RETURN_CODE(item1,item2)  (item1)
 #endif
